refactor(Application): seperate `Game` from `Application`
